+++ main.go
@@ -7,7 +7,6 @@ import (
 	"io/ioutil"
 	"os"
 	"reflect"
-	"sort"
 	"time"
 
 	"github.com/crowdsecurity/crowdsec/pkg/acquisition"
@@ -59,27 +58,6 @@ func getCmpOptions() cmp.Option {
 	return opt
 }
 
-//sort alert for each overflow
-func sortAlerts(event types.Event) types.Event {
-	//copy the slice
-	if event.Overflow.APIAlerts == nil {
-		return event
-	}
-	//	tmp := append([]models.Alert{}, event.Overflow.APIAlerts...)
-	//	log.Printf("meta1: %+v", event)
-	for index, alert := range event.Overflow.APIAlerts {
-		for i, evt := range alert.Events {
-			meta := evt.Meta
-			sort.Slice(meta, func(i, j int) bool {
-				return meta[i].Key < meta[j].Key
-			})
-			//			log.Printf("meta2: %+v", meta)
-			event.Overflow.APIAlerts[index].Events[i].Meta = meta
-		}
-	}
-	return event
-}
-
 //cleanForMatch : cleanup results from items that might change every run. We strip as well strictly equal results
 func cleanForMatch(in map[string]map[string]types.Event) map[string]map[string]types.Event {
 	for stage, val := range in {
@@ -90,102 +68,6 @@ func cleanForMatch(in map[string]map[string]types.Event) map[string]map[string]t
 		}
 	}
 	return in
-}
-
-func parsePoMatchLine(event types.Event, parserCTX *parser.UnixParserCtx, parserNodes []parser.Node) (bool, bool, types.Event, error) {
-	var (
-		err    error
-		parsed types.Event
-	)
-	//	oneResult := LineParseResult{}
-
-	oneResult := LineParseResult{}
-
-	if event.Type == types.LOG {
-		return true, false, types.Event{}, fmt.Errorf("event %+v is not an overflow", event)
-	}
-
-	parsed, err = parser.Parse(*parserCTX, event, parserNodes)
-	if err != nil {
-		return false, false, types.Event{}, fmt.Errorf("failed parsing : %v\n", err)
-	}
-	if parsed.Overflow.Reprocess {
-		log.Infof("Pouring buckets")
-		_, err = leaky.PourItemToHolders(parsed, holders, buckets)
-	}
-	if !parsed.Process {
-		return true, false, types.Event{}, fmt.Errorf("unparsed line %s", parsed.Line.Raw)
-	}
-
-	//marshal current result
-	oneResult.Line = parsed.Line.Raw
-	//we need to clean Line's timestamp
-	oneResult.ParserResults = cleanForMatch(parser.StageParseCache)
-
-	opt := getCmpOptions()
-
-	/*
-		Iterate over the list of expected results and try to find back
-	*/
-	AllPoResults = append(AllPoResults, oneResult)
-	matched := false
-	for idx, candidate := range AllPoExpected {
-		//not our line
-		if candidate.Line != event.Line.Raw {
-			continue
-		}
-		if cmp.Equal(candidate, oneResult, opt) {
-			matched = true
-			//we go an exact match
-			AllPoExpected = append(AllPoExpected[:idx], AllPoExpected[idx+1:]...)
-		} else {
-			return false, true, types.Event{}, fmt.Errorf("mismatch diff (-want +got) : %s", cmp.Diff(candidate, oneResult, opt))
-		}
-		break
-	}
-	if !matched && len(AllPoExpected) != 0 {
-		return false, true, types.Event{}, fmt.Errorf("Result is not in the %d expected results", len(AllPoExpected))
-	}
-	return matched, true, parsed, nil
-}
-
-func testBucketsOutput(target_dir string, AllBucketsResult []types.Event) (bool, error) {
-	var (
-		OrigExpectedLen    int
-		AllBucketsExpected []types.Event = []types.Event{}
-	)
-	//load the expected results
-	ExpectedPresent := false
-	expectedResultsFile := target_dir + "/buckets_results.json"
-	expected_bytes, err := ioutil.ReadFile(expectedResultsFile)
-	if err != nil {
-		log.Warningf("no buckets result in %s, will dump data instead!", target_dir)
-	} else {
-		if err := json.Unmarshal(expected_bytes, &AllBucketsExpected); err != nil {
-			return false, fmt.Errorf("file %s can't be unmarshaled : %s", expectedResultsFile, err)
-		} else {
-			ExpectedPresent = true
-			OrigExpectedLen = len(AllBucketsExpected)
-		}
-	}
-	//there was no data present, just dump
-	if !ExpectedPresent {
-		log.Warningf("No expected results loaded, dump.")
-		dump_bytes, err := json.MarshalIndent(AllBucketsResult, "", " ")
-		if err != nil {
-			log.Fatalf("failed to marshal results : %s", err)
-		}
-		if err := ioutil.WriteFile(expectedResultsFile, dump_bytes, 0644); err != nil {
-			log.Fatalf("failed to dump data to %s : %s", expectedResultsFile, err)
-		}
-	} else {
-		if len(AllExpected) > 0 {
-			log.Errorf("Left-over results in expected : %d", len(AllExpected))
-		}
-	}
-
-	return true, nil
-
 }
 
 func testOneDir(target_dir string, parsers *parser.Parsers, cConfig *csconfig.GlobalConfig) (bool, error) {
@@ -389,40 +271,6 @@ func testOneDir(target_dir string, parsers *parser.Parsers, cConfig *csconfig.Gl
 		log.Warningf("acquisition returned error : %s", err)
 	}
 
-	//from here we will deal with postoverflow
-	opt := getCmpOptions()
-	matched := false
-	if cmp.Equal(AllBucketsExpected, AllBucketsResult, opt) {
-		matched = true
-	} else {
-		expectedResultsFile = expectedResultsFile + ".fail"
-		log.Errorf("tests failed, writting results to %s", expectedResultsFile)
-		dump_bytes, err := json.MarshalIndent(AllBucketsResult, "", " ")
-		if err != nil {
-			log.Fatalf("failed to marshal results : %s", err)
-		}
-		if err := ioutil.WriteFile(expectedResultsFile, dump_bytes, 0644); err != nil {
-			log.Fatalf("failed to dump data to %s : %s", expectedResultsFile, err)
-		}
-		log.Printf("done")
-		return false, fmt.Errorf("mismatch diff (-want +got) : %s", cmp.Diff(AllBucketsExpected, AllBucketsResult, opt))
-	}
-
-	if !matched && len(AllExpected) != 0 {
-		expectedResultsFile = expectedResultsFile + ".fail"
-		log.Errorf("tests failed, writting results to %s", expectedResultsFile)
-		dump_bytes, err := json.MarshalIndent(AllBucketsResult, "", " ")
-		if err != nil {
-			log.Fatalf("failed to marshal results : %s", err)
-		}
-		if err := ioutil.WriteFile(expectedResultsFile, dump_bytes, 0644); err != nil {
-			log.Fatalf("failed to dump data to %s : %s", expectedResultsFile, err)
-		}
-		log.Printf("done")
-		return false, fmt.Errorf("Result is not in the %d expected results", len(AllExpected))
-
-	}
-	log.Infof("%d/%d matched results", OrigExpectedLen-len(AllBucketsExpected), OrigExpectedLen)
 	log.Infof("tests are finished.")
 
 	return true, nil
