+++ parser.go
@@ -1,3 +1,4 @@
+package main
 
 import (
 	"crypto/sha256"
